env_commands now norm conform and somewhat tested but as far as i have seen it still works
